Change to general solution for labels
