+++ apps/frontend/src/components/questionary/questionaryComponents/DatePicker/QuestionDateForm.tsx
@@ -1,7 +1,6 @@
 import DateAdapter from '@mui/lab/AdapterLuxon';
 import LocalizationProvider from '@mui/lab/LocalizationProvider';
 import useTheme from '@mui/material/styles/useTheme';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, TextField } from 'formik-mui';
 import { DatePicker, DateTimePicker } from 'formik-mui-lab';
@@ -15,6 +14,7 @@ import { DateConfig, SettingsId } from 'generated/sdk';
 import { minMaxDateTimeCalculations } from 'utils/Time';
 import { useNaturalKeySchema } from 'utils/userFieldValidationSchema';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import { QuestionFormShell } from '../QuestionFormShell';
 
 export const QuestionDateForm = (props: QuestionFormProps) => {
@@ -28,14 +28,6 @@ export const QuestionDateForm = (props: QuestionFormProps) => {
   const field = props.question;
 
   const naturalKeySchema = useNaturalKeySchema(field.naturalKey);
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
 
   return (
     <QuestionFormShell
@@ -89,9 +81,7 @@ export const QuestionDateForm = (props: QuestionFormProps) => {
 
         return (
           <>
-            <label className={classes.label}>
-              You are editing the question
-            </label>
+            <TemplateEdit pageType="Question" />
             <Field
               name="naturalKey"
               label="Key"
