+++ apps/frontend/src/components/questionary/questionaryComponents/Embellishment/QuestionEmbellishmentForm.tsx
@@ -1,4 +1,3 @@
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, TextField } from 'formik-mui';
 import React from 'react';
@@ -9,6 +8,7 @@ import { QuestionFormProps } from 'components/questionary/QuestionaryComponentRe
 import { EmbellishmentConfig } from 'generated/sdk';
 import { useNaturalKeySchema } from 'utils/userFieldValidationSchema';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import { QuestionFormShell } from '../QuestionFormShell';
 
 export const QuestionEmbellishmentForm = (props: QuestionFormProps) => {
@@ -17,14 +17,6 @@ export const QuestionEmbellishmentForm = (props: QuestionFormProps) => {
   document.addEventListener('focusin', function (e) {
     e.stopImmediatePropagation();
   });
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
 
   return (
     <QuestionFormShell
@@ -39,7 +31,7 @@ export const QuestionEmbellishmentForm = (props: QuestionFormProps) => {
     >
       {(formikProps) => (
         <>
-          <label className={classes.label}>You are editing the question</label>
+          <TemplateEdit pageType="Question" />
           <Field
             name="naturalKey"
             label="Key"
