+++ apps/frontend/src/components/questionary/questionaryComponents/FileUpload/QuestionFileUploadForm.tsx
@@ -1,5 +1,4 @@
 import ListItemText from '@mui/material/ListItemText';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, Select, TextField } from 'formik-mui';
 import React from 'react';
@@ -11,6 +10,7 @@ import { QuestionFormProps } from 'components/questionary/QuestionaryComponentRe
 import { FileUploadConfig } from 'generated/sdk';
 import { useNaturalKeySchema } from 'utils/userFieldValidationSchema';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import { QuestionFormShell } from '../QuestionFormShell';
 
 export const QuestionFileUploadForm = (props: QuestionFormProps) => {
@@ -24,14 +24,6 @@ export const QuestionFileUploadForm = (props: QuestionFormProps) => {
     { label: 'video/*', value: 'video/*' },
     { label: 'image/*', value: 'image/*' },
   ];
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
 
   return (
     <QuestionFormShell
@@ -59,9 +51,7 @@ export const QuestionFileUploadForm = (props: QuestionFormProps) => {
 
         return (
           <>
-            <label className={classes.label}>
-              You are editing the question
-            </label>
+            <TemplateEdit pageType="Question" />
             <Field
               name="naturalKey"
               id="Key-Input"
