+++ apps/frontend/src/components/questionary/questionaryComponents/FileUpload/QuestionTemplateRelationFileUploadForm.tsx
@@ -1,5 +1,4 @@
 import ListItemText from '@mui/material/ListItemText';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, Select, TextField } from 'formik-mui';
 import React from 'react';
@@ -10,6 +9,7 @@ import TitledContainer from 'components/common/TitledContainer';
 import { QuestionTemplateRelationFormProps } from 'components/questionary/QuestionaryComponentRegistry';
 import { FileUploadConfig } from 'generated/sdk';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import QuestionDependencyList from '../QuestionDependencyList';
 import { QuestionExcerpt } from '../QuestionExcerpt';
 import { QuestionTemplateRelationFormShell } from '../QuestionTemplateRelationFormShell';
@@ -25,14 +25,6 @@ export const QuestionTemplateRelationFileUploadForm = (
     { label: 'video/*', value: 'video/*' },
     { label: 'image/*', value: 'image/*' },
   ];
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
 
   return (
     <QuestionTemplateRelationFormShell
@@ -59,10 +51,7 @@ export const QuestionTemplateRelationFileUploadForm = (
 
           return (
             <>
-              <label className={classes.label}>
-                You are editing the question as it appears on the current
-                template
-              </label>
+              <TemplateEdit pageType="Template" />
               <QuestionExcerpt question={props.questionRel.question} />
               <TitledContainer label="Options">
                 <Field
