+++ apps/frontend/src/components/questionary/questionaryComponents/GenericTemplate/QuestionGenericTemplateForm.tsx
@@ -1,6 +1,5 @@
 import { Collapse, FormControl } from '@mui/material';
 import Link from '@mui/material/Link';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, TextField } from 'formik-mui';
 import React from 'react';
@@ -13,6 +12,7 @@ import { SubTemplateConfig, TemplateGroupId } from 'generated/sdk';
 import { useActiveTemplates } from 'hooks/call/useCallTemplates';
 import { useNaturalKeySchema } from 'utils/userFieldValidationSchema';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import { QuestionFormShell } from '../QuestionFormShell';
 
 export const QuestionGenericTemplateForm = (props: QuestionFormProps) => {
@@ -20,14 +20,6 @@ export const QuestionGenericTemplateForm = (props: QuestionFormProps) => {
   const naturalKeySchema = useNaturalKeySchema(field.naturalKey);
 
   const { templates } = useActiveTemplates(TemplateGroupId.GENERIC_TEMPLATE);
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
   if (!templates) {
     return null;
   }
@@ -57,7 +49,7 @@ export const QuestionGenericTemplateForm = (props: QuestionFormProps) => {
     >
       {(formikProps) => (
         <>
-          <label className={classes.label}>You are editing the question</label>
+          <TemplateEdit pageType="Question" />
           <Field
             name="naturalKey"
             id="Key-Input"
