+++ apps/frontend/src/components/questionary/questionaryComponents/InstrumentPicker/QuestionTemplateRelationInstrumentPickerForm.tsx
@@ -1,26 +1,16 @@
-import makeStyles from '@mui/styles/makeStyles';
 import React from 'react';
 import * as Yup from 'yup';
 
 import { QuestionTemplateRelationFormProps } from 'components/questionary/QuestionaryComponentRegistry';
 import { InstrumentPickerConfig } from 'generated/sdk';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import { QuestionExcerpt } from '../QuestionExcerpt';
 import { QuestionTemplateRelationFormShell } from '../QuestionTemplateRelationFormShell';
 import { QuestionInstrumentPickerFormCommon } from './QuestionInstrumentPickerFormCommon';
-
 export const QuestionTemplateRelationInstrumentPickerForm = (
   props: QuestionTemplateRelationFormProps
 ) => {
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
-
   return (
     <QuestionTemplateRelationFormShell
       {...props}
@@ -33,9 +23,7 @@ export const QuestionTemplateRelationInstrumentPickerForm = (
     >
       {() => (
         <>
-          <label className={classes.label}>
-            You are editing the question as it appears on the current template
-          </label>
+          <TemplateEdit pageType="Template" />
           <QuestionExcerpt question={props.questionRel.question} />
           <QuestionInstrumentPickerFormCommon
             config={props.questionRel.question.config as InstrumentPickerConfig}
