+++ apps/frontend/src/components/questionary/questionaryComponents/Interval/QuestionIntervalForm.tsx
@@ -1,6 +1,5 @@
 import Autocomplete from '@mui/lab/Autocomplete';
 import MaterialTextField from '@mui/material/TextField';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, TextField } from 'formik-mui';
 import React, { useState } from 'react';
@@ -13,20 +12,13 @@ import { IntervalConfig } from 'generated/sdk';
 import { useUnitsData } from 'hooks/settings/useUnitData';
 import { useNaturalKeySchema } from 'utils/userFieldValidationSchema';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 export const QuestionIntervalForm = (props: QuestionFormProps) => {
   const field = props.question;
   const intervalConfig = props.question.config as IntervalConfig;
   const naturalKeySchema = useNaturalKeySchema(field.naturalKey);
   const { units } = useUnitsData();
   const [selectedUnits, setSelectedUnits] = useState(intervalConfig.units);
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
 
   return (
     <QuestionFormShell
@@ -50,7 +42,7 @@ export const QuestionIntervalForm = (props: QuestionFormProps) => {
     >
       {({ setFieldValue }) => (
         <>
-          <label className={classes.label}>You are editing the question</label>
+          <TemplateEdit pageType="Question" />
           <Field
             name="naturalKey"
             label="Key"
