+++ apps/frontend/src/components/questionary/questionaryComponents/NumberInput/QuestionTemplateRelationNumberInputForm.tsx
@@ -1,6 +1,5 @@
 import Autocomplete from '@mui/lab/Autocomplete';
 import MaterialTextField from '@mui/material/TextField';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field, getIn } from 'formik';
 import { CheckboxWithLabel, TextField } from 'formik-mui';
 import React, { useState } from 'react';
@@ -13,6 +12,7 @@ import { QuestionExcerpt } from 'components/questionary/questionaryComponents/Qu
 import { NumberInputConfig, NumberValueConstraint } from 'generated/sdk';
 import { useUnitsData } from 'hooks/settings/useUnitData';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import QuestionDependencyList from '../QuestionDependencyList';
 import { QuestionTemplateRelationFormShell } from '../QuestionTemplateRelationFormShell';
 
@@ -22,14 +22,6 @@ export const QuestionTemplateRelationNumberForm = (
   const numberConfig = props.questionRel.config as NumberInputConfig;
   const { units } = useUnitsData();
   const [selectedUnits, setSelectedUnits] = useState(numberConfig.units);
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
 
   return (
     <QuestionTemplateRelationFormShell
@@ -59,9 +51,7 @@ export const QuestionTemplateRelationNumberForm = (
 
         return (
           <>
-            <label className={classes.label}>
-              You are editing the question as it appears on the current template
-            </label>
+            <TemplateEdit pageType="Template" />
             <QuestionExcerpt question={props.questionRel.question} />
             <Field
               name="config.small_label"
