+++ apps/frontend/src/components/questionary/questionaryComponents/SampleDeclaration/QuestionSampleDeclarationForm.tsx
@@ -1,6 +1,5 @@
 import FormControl from '@mui/material/FormControl';
 import Link from '@mui/material/Link';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { TextField } from 'formik-mui';
 import React, { useContext, ChangeEvent } from 'react';
@@ -19,6 +18,7 @@ import {
 import { useActiveTemplates } from 'hooks/call/useCallTemplates';
 import { useNaturalKeySchema } from 'utils/userFieldValidationSchema';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import { QuestionFormShell } from '../QuestionFormShell';
 
 export const QuestionSampleDeclarationForm = (props: QuestionFormProps) => {
@@ -30,14 +30,6 @@ export const QuestionSampleDeclarationForm = (props: QuestionFormProps) => {
     templates: sampleTemplates,
     refreshTemplates: refreshSampleTemplates,
   } = useActiveTemplates(TemplateGroupId.SAMPLE, config.templateId);
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
   const { templates: esiTemplates, refreshTemplates: refreshEsiTemplates } =
     useActiveTemplates(TemplateGroupId.SAMPLE_ESI, config.esiTemplateId);
 
@@ -70,7 +62,7 @@ export const QuestionSampleDeclarationForm = (props: QuestionFormProps) => {
     >
       {({ values, setFieldValue }) => (
         <>
-          <label className={classes.label}>You are editing the question</label>
+          <TemplateEdit pageType="Question" />
           <Field
             name="naturalKey"
             id="Key-Input"
