+++ apps/frontend/src/components/questionary/questionaryComponents/TextInput/QuestionTemplateRelationTextInputForm.tsx
@@ -1,6 +1,5 @@
 import Box from '@mui/material/Box';
 import Collapse from '@mui/material/Collapse';
-import makeStyles from '@mui/styles/makeStyles';
 import { Field } from 'formik';
 import { CheckboxWithLabel, TextField } from 'formik-mui';
 import React, { ChangeEvent } from 'react';
@@ -11,22 +10,13 @@ import TitledContainer from 'components/common/TitledContainer';
 import { QuestionTemplateRelationFormProps } from 'components/questionary/QuestionaryComponentRegistry';
 import { TextInputConfig } from 'generated/sdk';
 
+import TemplateEdit from '../../../template/QuestionTemplateLabel';
 import QuestionDependencyList from '../QuestionDependencyList';
 import { QuestionExcerpt } from '../QuestionExcerpt';
 import { QuestionTemplateRelationFormShell } from '../QuestionTemplateRelationFormShell';
-
 export const QuestionTemplateRelationTextInputForm = (
   props: QuestionTemplateRelationFormProps
 ) => {
-  const useStyles = makeStyles((theme) => ({
-    label: {
-      color: theme.palette.primary.main,
-      backgroundColor: theme.palette.grey[300],
-      fontSize: 'medium',
-    },
-  }));
-  const classes = useStyles();
-
   return (
     <QuestionTemplateRelationFormShell
       {...props}
@@ -43,9 +33,7 @@ export const QuestionTemplateRelationTextInputForm = (
     >
       {(formikProps) => (
         <>
-          <label className={classes.label}>
-            You are editing the question as it appears on the current template
-          </label>
+          <TemplateEdit pageType="Template" />
           <QuestionExcerpt question={props.questionRel.question} />
 
           <TitledContainer label="Constraints">
